requirements.txt
altair==5.0.1
ansi2html==1.8.0
anyio==3.7.1
anytree==2.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
category-encoders==2.5.1.post0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.1.0
cycler==0.11.0
cymem==2.0.7
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.41.1
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
hdbcli==2.17.22
htbuilder==0.6.2
ibm-db==3.1.4
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.24.0
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpickle==3.0.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook_shim==0.2.3
numpy==1.25.0
openpyxl==3.0.10
overrides==7.3.1
packaging==23.1
pandarallel==1.6.5
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
patsy==0.5.3
pickleshare==0.7.5
pikepdf==8.1.1
Pillow==9.5.0
platformdirs==3.8.1
Plost==0.2.5
plotly==5.15.0
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
#pywin32==306
#pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlalchemy-hana==0.5.0
srsly==2.4.7
stack-data==0.6.2
streamlit==1.24.0
streamlit_authenticator==0.2.3
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
treelib==1.6.4
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.0.3
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
wordcloud==1.9.2
yellowbrick==1.5
zipp==3.15.0
